(i) cart system changed so that users are limited to 20 quantity per order (ii) proper alert message when interacting with cart (iii) proper alert message when trader manages, add or updates product (CRUD), proper alert message when user interacts with profile or their dashboard and other minor changes related to session here and there

+++ delete.php
@@ -14,9 +14,19 @@
 
 		if (isset($_GET['ID'])) 
 		{
-			$delete_item = "Delete from cart where ID = ".$_GET['ID'];
+			$delete_item = "Delete from cart where ID = ".$_GET['ID']. "AND FK1_USER_ID=" .$_SESSION['id'];
 			$delete_result = oci_parse($conn, $delete_item);
 			oci_execute($delete_result);
+			$_SESSION['passmessage'] = "Item deleted successfully";
+			header("Location: temp.php");
+		}
+
+		if (isset($_POST['deleteBtn'])) 
+		{
+			$delete_item = "Delete from cart WHERE FK1_USER_ID=" .$_SESSION['id'];
+			$delete_result = oci_parse($conn, $delete_item);
+			oci_execute($delete_result);
+			$_SESSION['passmessage'] = "Cart emptied successfully";
 			header("Location: temp.php");
 		}
